requirements.txt
asgiref==3.3.4
Django==3.2.3
django-swagger-render==0.1.1
Markdown==3.3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
django-dirtyfields~=1.7.0
celery~=3.1.26.post2